requirements.txt
djangorestframework==3.8.2
Django==2.1.5
Pillow==5.4.1
